http://www.adenforshaw.co.uk/?p=219 An easy to use AS3 shape drawing class, gear, cogs, wedge, arc, star, burst
This is mainly just an Actionscript 3 implementation of the hard work done by both Kevin Williams and Ric Ewing. I’ve also used a Minimal Comps Slider as well and included the slightly amended code for that which does not require the Flex SDK (makes it easier for people not familiar with it).